+++ packages/core/webpack/src/plugins/MoveFilesIntoDistFolderPlugin.ts
@@ -1,4 +1,4 @@
-import { Compiler } from "webpack";
+import { type Compiler } from "webpack";
 import { promises as fs } from "fs";
 import path from "path";