+++ packages/react/src/atomic/Link/src/Link.tsx
@@ -1,18 +1,18 @@
 import React, {
-  ComponentPropsWithRef,
+  type ComponentPropsWithRef,
   forwardRef,
   useEffect,
   useImperativeHandle,
   useRef,
-  ForwardedRef,
+  type ForwardedRef,
 } from "react";
 import {
-  PolymorphicForwardRefComponent,
+  type PolymorphicForwardRefComponent,
   useRefObjectAsForwardedRef,
 } from "@nimbus-ds/typings";
 import { link } from "@nimbus-ds/styles";
 
-import { LinkBaseProps, LinkComponents } from "./link.types";
+import { type LinkBaseProps, type LinkComponents } from "./link.types";
 import { LinkSkeleton } from "./components";
 
 const Link = forwardRef(