+++ packages/react/src/common/storybook/sampleBox.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Box, BoxProps } from "@nimbus-ds/box";
+import { Box, type BoxProps } from "@nimbus-ds/box";
 
 /**
  * A styled container component for Storybook examples that provides