+++ packages/react/src/composite/Stepper/src/stepper.spec.tsx
@@ -3,8 +3,8 @@ import { render, screen, fireEvent } from "@testing-library/react";
 
 import { Stepper } from "./Stepper";
 import {
-  BaseStepperProperties,
-  ControlledStepperProperties,
+  type BaseStepperProperties,
+  type ControlledStepperProperties,
 } from "./stepper.types";
 
 const makeUncontrolledSut = (props: BaseStepperProperties) => {
@@ -48,7 +48,9 @@ describe("GIVEN <Stepper />", () => {
       expect(stepper.getAttribute("aria-valuenow")).toBe("2");
       expect(stepper.getAttribute("aria-valuemin")).toBe("1");
       expect(stepper.getAttribute("aria-valuemax")).toBe("3");
-      expect(stepper.getAttribute("aria-label")).toBe("Multi-step process: Step 2 of 3");
+      expect(stepper.getAttribute("aria-label")).toBe(
+        "Multi-step process: Step 2 of 3"
+      );
     });
 
     it("THEN should render with single step", () => {
@@ -73,7 +75,9 @@ describe("GIVEN <Stepper />", () => {
 
       const stepper = screen.getByTestId("stepper-element");
       expect(stepper.getAttribute("aria-valuenow")).toBe("1");
-      expect(stepper.getAttribute("aria-label")).toBe("Multi-step process: Step 1 of 2");
+      expect(stepper.getAttribute("aria-label")).toBe(
+        "Multi-step process: Step 1 of 2"
+      );
     });
 
     it("THEN should handle maximum activeStep", () => {