+++ packages/react/src/composite/Table/src/components/TableBody/tableBody.types.ts
@@ -1,4 +1,4 @@
-import { ReactNode, HTMLAttributes } from "react";
+import { type ReactNode, type HTMLAttributes } from "react";
 
 export interface TableBodyProperties {
   /**