+++ packages/react/src/composite/Tabs/src/components/TabsButton/tabsButton.spec.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { fireEvent, render, screen } from "@testing-library/react";
 
 import { TabsButton } from "./TabsButton";
-import { TabsButtonProps } from "./tabsButton.types";
+import { type TabsButtonProps } from "./tabsButton.types";
 
 const buttonLabel = "myButtonLabel";
 const buttonLabelReactNode = <p data-testid="label-element">myButtonContent</p>;